Gerenalize the input stream processors accept
So far the stream processors can only process streams in the guise of
non-well-founded lists. For real-world streams this seems too
restrictive and it seems more appropriate to allow any type which is
an instance of the stream-interface as input type for stream
processor. So the input type is replaced by a generic contraint by the
according trait bound.