Add complementary stream implementation
There is a new submodule of streams defining a variant of
message-receivers suited to implement the stream-trait. The
accompanying implementation of the stream-trait is particularly useful
for stream processors operating on event-streams as it allows to
construct streams 'event by event' and not only corecursively like in
the case of infinite lists.