Relax lifetime constraints
By default trait objects enforce an unbounded lifetime on the references
they contain. This also holds for the trait objects in the stream
processors and infinite lists of rspl, of course. Relaxing this makes
the library more flexible at the cost of a slightly more complex API.
But as one can think of use cases requiring that flexibility it is
worth the price.