Relax trait bounds
Both streams and inputs to stream processors require a Copy-bound so
far. This can easily be relaxed without doing any severe harm as the
Copy-trait naturally embeds in the Clone-trait and one has to put some
effort into destroying the embedding. Doing so also extends the
application domain of rspl. Hence it is a good idea to incorporate the
relaxation.