Refine stream processor semantic
When a get-processor is evaluated on a stream the recursive call is
always on the tail of that stream. This is ok if the stream does already
exist as a whole (if it was defined corecursively or so). But it is bad
if the infinity-property of the stream originates from the
infinity-property of time because then evaluation might block on
put-processing making rspl inapplicable for reactive programming. But
allowing for reactive programming is one of the main purposes of rspl.
So the problem needs a solution: the semantic of stream processors is
refined such that the tail is only taken when it is really needed -
namely only if the next evaluation step is get-processing.

(Note that the change is local in the sense that it does only change
the behavior of stream processors which process not already existing
input. Moreover the change does not affect what is outputted but only
when it is outputted.)