Tail-recursion-eliminate rspl manually
Rust does not always do tail-recursion-elimination making
stack-overflows more likely. As the application domain of rspl includes
event-driven programming where programs are supposed to run for long
periods of time the implementation should do everything in its power to
prevent stack-overflows. In this case it means to incorporate the
optimization into the source code.