Add optional no-std support
rspl essentially works without the standard library if there is
something like a box of the alloc-crate. As event-driven programming
with state machines - a possible application domain of rspl - seems
common in embedded settings, too, it makes sense to offer a no-std
option, at least. The commit achieves this with the standard technique.