Edit keywords and categories in manifest
Keywords: Reflecting the "event-" and "demand-driven" aspects of rspl
corresponding to its language constructs in the keyword list is better
than just listing "stream" and "events". This is due to symmetry and
because "stream" is already contained in "stream-processing".
Categories: On the one hand rspl is not primarily for "embedded" use.
So this category is better weakened to "no-std". On the other hand rspl
seems to fulfill the "asynchronous" and "concurrency"-definition in
some respect and it seems reasonble to add them. Moreover, since rspl
is a language, the "compiler" category is a must.