Fix semantic bug in `Put`-wrapper
The semantic of `put` does not quite wrap that of `Put`. The problem is
that when calling `put` on a computation returning a stream processor
that computation is eagerly evaluated whereas with `Put` it would be an
unevaluated thunk which is only evaluated later. This is not what users
expect and therefore it is changed.