Streamline tests and examples
For both event- and demand-driven there is an abstract usage-examples
and a full-blown concrete example as well as an integration test. As
event and demand-driven is somewhat dual and since the example have a
relation to the tests it makes sense to keep the deployed
implementation-patterns in sync.
This commit can be considered the final touch to achieve that goal.