Replace Hancock reference
The assumption that Hancock is the inventor of the language implemented
here is wrong. It already appears in earlier work by Hughes (and seems
to have its origin in the PhD-thesis introducing fudgets). The
discussion in Hughes's work also seems nicer than in the Hancock
reference. So it is better to replace the Hancock reference by the
Hughes reference.