Fix semantic bug in `Cons`-wrapper
The semantic of `cons` does not quite wrap that of `Cons`. The problem
is that when calling `cons` on a computation returning an infinite list
that computation is eagerly evaluated whereas with `Cons` it would be
an unevaluated thunk which is only evaluated later. This is not what
users expect and therefore it is changed.